# backend/src/main/java/com/ups/glam/backend/domain/comment/CreateCommentRequest.java
package com.ups.glam.backend.domain.comment;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class CreateCommentRequest {
    @NotBlank
    private String content;
}